+++ src/main/java/service/LottoPrizeCounter.java
@@ -1,6 +1,7 @@
 package service;
 
 import domain.lotto.LottoPrize;
+import domain.lotto.LottoPrizeResult;
 import domain.lotto.LottoTicket;
 import domain.lotto.LottoWinningNumber;
