+++ src/test/java/domain/lotto/LottoPrizeResultTest.java
@@ -1,4 +1,4 @@
-package service;
+package domain.lotto;
 
 import domain.lotto.*;
 import org.junit.jupiter.api.BeforeEach;
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import service.LottoPrizeResult;
+import domain.lotto.LottoPrizeResult;
 import service.LottoShop;
 
 import java.util.HashMap;
